Move LocalDataset to torch package
