+++ darwin/dataset/__init__.py
@@ -1,2 +1,1 @@
-from darwin.dataset.local_dataset import LocalDataset  # noqa
 from darwin.dataset.remote_dataset import RemoteDataset  # noqa
